requirements.txt
torch
torchvision
Pillow
python-dotenv
tqdm
